Add files via upload
Welcome to the repository of my final project as part of the Blockchain Expert Course by Deusto and UCAM!

In this project, I've utilized Solidity to develop a Smart Contract designed for an Initial Coin Offering (ICO). This project not only represents the culmination of my blockchain studies but also demonstrates my ability to apply this knowledge to real-world industry use cases.

Throughout the development of this Smart Contract, I've gained solid skills in Solidity programming, as well as a deep understanding of the fundamental principles behind ICOs and smart contracts. Furthermore, I've explored and implemented best practices for security and efficiency to ensure the integrity and functionality of the contract.

This project not only showcases my technical ability but also my capacity to apply theoretical concepts in practical and challenging environments. I'm excited to share this work with the community and to explore opportunities to continue growing and contributing to the exciting world of blockchain technology.

Thank you for reviewing my project and considering my profile as a blockchain developer!