Add files via upload
Project Overview: CoffeeToken Sale
This project is a smart contract implementation for a token sale on the Ethereum blockchain. The project consists of two primary Solidity files:

CoffeeToken Contract (coffeeTokenSale.sol): This contract is designed to mint and manage the "CoffeeToken," an ERC20 token used for purchasing coffee in a decentralized manner.
ERC20 Token Sale Contract (ERC20_Token_Sale.sol): This contract handles the sale of the CoffeeToken, allowing users to buy tokens with Ether.
The overall goal of this project is to create a decentralized application (DApp) where users can mint, purchase, and transfer CoffeeTokens.